+++ resources/lang/de/people.php
@@ -218,4 +218,7 @@ return [
     'validity_in_days' => 'Gültigkeit in Tagen',
     'file_must_be_excel_cvs_containing_columns_name_position' => 'Datei muss im Excel oder CSV Format vorliegen und die Spalten \'Name\', \'Position\' und optional \'Code\' beinhalten.',
     'staff_card_no' => 'Mitarbeiterkartennummer',
+    'column_set' => 'Spalten',
+    'name_nationality_occupation' => 'Nationalität, Beschäftigung',
+    'contact_info' => 'Kontaktinformationen',
 ];