+++ resources/lang/en/people.php
@@ -218,4 +218,7 @@ return [
     'validity_in_days' => 'Validity in days',
     'file_must_be_excel_cvs_containing_columns_name_position' => 'File must be in Excel or CSV format and contain the columns \'Name\', \'Position\' and optional \'Code\'.',
     'staff_card_no' => 'Staff card number',
+    'column_set' => 'Columns',
+    'name_nationality_occupation' => 'Nationality, Occupation',
+    'contact_info' => 'Contact information',
 ];